feat(test): Refactor
