+++ cg/apps/mip/parse_trending.py
@@ -1,5 +1,3 @@
-from typing import List
-
 from cg.models.mip.mip_config import MipBaseConfig
 from cg.models.mip.mip_metrics_deliverables import MetricsDeliverables
 from cg.models.mip.mip_sample_info import MipBaseSampleInfo
@@ -17,22 +15,16 @@ def parse_mip_analysis(mip_config_raw: dict, qcmetrics_raw: dict, sampleinfo_raw
         dict: parsed data
     """
     mip_config: MipBaseConfig = MipBaseConfig(**mip_config_raw)
-    sample_info: MipBaseSampleInfo = MipBaseSampleInfo(**sampleinfo_raw)
     qc_metrics = MetricsDeliverables(**qcmetrics_raw)
+    sample_info: MipBaseSampleInfo = MipBaseSampleInfo(**sampleinfo_raw)
+
     outdata = {
         "id_metrics": qc_metrics.id_metrics,
         "case": mip_config.case_id,
         "genome_build": sample_info.genome_build,
         "mip_version": sample_info.mip_version,
         "rank_model_version": sample_info.rank_model_version,
-        "sample_ids": _get_sample_ids(mip_config=mip_config),
+        "sample_ids": mip_config.sample_ids,
         "sv_rank_model_version": sample_info.sv_rank_model_version,
     }
     return outdata
-
-
-def _get_sample_ids(mip_config: MipBaseConfig) -> List:
-    sample_ids: list = []
-    for sample_data in mip_config.samples:
-        sample_ids.append(sample_data.sample_id)
-    return sample_ids
